refactor(data): Repository 레이어 단순화에 따른 불필요한 인터페이스 및 인프라 구현체 삭제
