--- Data/Sources/Infrastructure/Summary/SummaryGenerationResult.swift
@@ -1,9 +1,0 @@
-#if canImport(FoundationModels)
-    import FoundationModels
-
-    @Generable
-    struct SummaryGenerationResult {
-        let keywords: [String]
-        let summary: String
-    }
-#endif
