--- Data/Sources/Repositories/VoiceNotes/STTRepositoryError+Mapping.swift
@@ -1,14 +1,0 @@
-import Domain
-
-extension STTRepositoryError {
-    init(_ error: STTServiceError) {
-        switch error {
-        case .cancelled:
-            self = .cancelled
-        case .alreadyTranscribing, .recognizerUnavailable, .transcribeFailed:
-            self = .transcribeFailed
-        case .unknown(let error):
-            self = .unknown(error)
-        }
-    }
-}
